+++ src/CAServer.Grains/Grain/RedPackage/IRedPackageGrain.cs
@@ -5,7 +5,7 @@ namespace CAServer.Grains.Grain.RedPackage;
 
 public interface IRedPackageGrain : IGrainWithGuidKey
 {
-    Task<GrainResultDto<RedPackageDetailDto>> CreateRedPackage(SendRedPackageInputDto input, int decimalIn, decimal minAmount,
+    Task<GrainResultDto<RedPackageDetailDto>> CreateRedPackage(SendRedPackageInputDto input, int decimalIn, long minAmount,
         Guid senderId);
     Task<GrainResultDto<RedPackageDetailDto>> GetRedPackage(int skip, int max,Guid userId);
     Task<GrainResultDto<bool>> IsUserIdGrab(Guid userId);
